refactor to client and server package
