+++ Sudoku Smasher/src/SmasherClient/ClientMain.java
@@ -1,0 +1,10 @@
+package SmasherClient;
+
+import SmasherClient.Frame;
+public class ClientMain {
+    public static void main(String[] args)
+    {
+        Frame.createAndShowGUI();
+        Client client = new Client();
+    }
+}
