+++ Sudoku Smasher/src/SmasherClient/Frame.java
@@ -1,4 +1,4 @@
-package SmasherServer;
+package SmasherClient;
 
 import javax.swing.*;
 import javax.swing.border.Border;
